put back tests and fixed them for current implementation

+++ FootballLeague/Controllers/MatchesController.cs
@@ -52,6 +52,9 @@ namespace FootballLeague.Controllers
             if (match == null)
                 return;
 
+            if (!_matchRepository.MatchContainsTeam(match, teamId))
+                return;
+
             if (match.Contains(user))
                 _matchRepository.RemoveMatchParticipantFromTeam(user, match, teamId);
             else
